+++ CHANGELOG.md
@@ -8,9 +8,9 @@
 
 ### Fixed
 
+- Allow non-colliding extension leaves beneath core command groups ([pm-bwkmp4](https://github.com/unbraind/pm-cli/blob/main/.agents/pm/issues/pm-bwkmp4.toon))
 - Compact default list receipts and avoid repeated recovery argument echoes ([pm-4f86c4](https://github.com/unbraind/pm-cli/blob/main/.agents/pm/issues/pm-4f86c4.toon))
 - Canonicalize unknown-command recovery with replacement flags ([pm-zsic8h](https://github.com/unbraind/pm-cli/blob/main/.agents/pm/issues/pm-zsic8h.toon))
-- Allow non-colliding extension leaves beneath core command groups ([pm-bwkmp4](https://github.com/unbraind/pm-cli/blob/main/.agents/pm/issues/pm-bwkmp4.toon))
 
 ## 2026.9.13 - 2026-09-13